GEODE-8609: Create a dunit suspect file per VM (#5625)
- Create a new `dunit_suspect` log for each VM. The logs will now
  be named `dunit_suspect-vm<VMID>.log`. The locator VM and the test
  runner VM will have logs named `dunit_suspect-locator.log` and
  `dunit_suspect-local.log` respectively.
- Leaving in the bad regexp which will be addressed by a subsequent Jira

+++ geode-dunit/src/main/java/org/apache/geode/test/dunit/internal/StandAloneDUnitEnv.java
@@ -51,7 +51,7 @@ public class StandAloneDUnitEnv extends DUnitEnv {
 
   @Override
   public int getId() {
-    return Integer.getInteger(DUnitLauncher.VM_NUM_PARAM, -1).intValue();
+    return Integer.getInteger(DUnitLauncher.VM_NUM_PARAM, -1);
   }
 
   @Override
